+++ src/pages/Login.js
@@ -33,8 +33,8 @@ export default class Login extends Component {
   };
 
   render() {
-    let { from } = this.props.location.state || { from: { pathname: '/' } };
-    let { redirectToReferrer } = this.state;
+    const { from } = this.props.location.state || { from: { pathname: '/' } };
+    const { redirectToReferrer } = this.state;
 
     if (redirectToReferrer) return <Redirect to={from} />;
