+++ src/styles/Link.scss
@@ -1,0 +1,3 @@
+.link--no-underline {
+  text-decoration: none;
+}
